wip - setting up integration test & helpers
