+++ db/index.js
@@ -1,5 +1,9 @@
 const { Client } = require('pg');
-var connectionString = 'postgres://alizeh:password@localhost:5432/questions_answers';
+require('dotenv').config();
+const username = process.env.POSTGRES_USERNAME;
+const password = process.env.POSTGRES_PASSWORD;
+
+var connectionString = `postgres://${username}:${password}@localhost:5432/questions_answers`;
 
 const client = new Client({
     connectionString: connectionString
