+++ package.json
@@ -21,10 +21,12 @@
   },
   "homepage": "https://github.com/SDC-Apricot/Questions_Answers_API#readme",
   "devDependencies": {
+    "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "jest": "^27.0.6",
     "nodemon": "^2.0.11",
     "pg": "^8.6.0",
-    "pg-hstore": "^2.3.4"
+    "pg-hstore": "^2.3.4",
+    "supertest": "^6.1.3"
   }
 }
